Update and rename simulation.R to interaction_bad.R
This is an example of getting simulated marginal effects for an interaction term. The approach here would be valid for a linear model, but (according to Ai and Norton) it gives inaccurate marginal effects in a non-linear model. Long story short: don't use this, it's just here for pedagogical purposes.